+++ .vscode/settings.json
@@ -1,5 +1,6 @@
 {
     "files.associations": {
-        "type_traits": "c"
+        "type_traits": "c",
+        "minishell.h": "c"
     }
 }
