+++ include/structs.h
@@ -6,7 +6,7 @@
 /*   By: myakoven <myakoven@student.42berlin.de>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/07/13 20:00:21 by myakoven          #+#    #+#             */
-/*   Updated: 2024/10/25 11:56:16 by myakoven         ###   ########.fr       */
+/*   Updated: 2024/10/25 15:38:23 by myakoven         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -66,6 +66,8 @@ typedef struct s_cmd		t_cmd;
 
 typedef struct s_tools
 {
+	struct sigaction		sa;
+
 	char					**env;
 	int						env_len;
 
@@ -81,7 +83,7 @@ typedef struct s_tools
 	struct s_cmd			*tree;
 
 	// pid_t					children[MAXARGS];
-	
+
 	/* Temporary */
 	char					*s;
 	char					*cmd_end;
