+++ main.c
@@ -6,27 +6,29 @@
 /*   By: myakoven <myakoven@student.42berlin.de>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/07/07 20:51:01 by myakoven          #+#    #+#             */
-/*   Updated: 2024/10/24 19:13:00 by myakoven         ###   ########.fr       */
+/*   Updated: 2024/10/25 15:41:33 by myakoven         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
+#include <signal.h>
 #include "./include/minishell.h"
 
-volatile sig_atomic_t	global_signal = 0;
+// volatile sig_atomic_t	global_signal = 0; // delete?
 
 int	main(int argc, char **argv, char **env)
 {
 	t_tools	tools;
 
-	// struct sigaction	sa;
+	struct sigaction	sa;
 	if (argc > 1 || argv[1])
 		ft_putstr_fd("This program does not accept arguments\n", 2);
+	
 	ft_memset(&tools, 0, sizeof(t_tools)); // init tools to zero
 	here_init(tools.heredocs, &tools);
 	copy_env(&tools, env);
 	if (!tools.env || !tools.heredocs[0][0])
 		(error_exit(&tools, 1));
-	// init_sa(&sa);
+	init_sa(&sa);
 	shell_loop(&tools);
 	return (0);
 }
@@ -35,8 +37,8 @@ int	shell_loop(t_tools *tools)
 {
 	while (1)
 	{
-		// if (global_signal == SIGTERM) // TODO? or done
-		// 	break ;
+		if (global_signal == SIGTERM) // TODO? or done
+			break ;
 		tools->line = readline("minishell: ");
 		global_signal = 0;
 		if (!valid_line(tools->line))
@@ -67,6 +69,29 @@ int	shell_loop(t_tools *tools)
 	return (0);
 }
 
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
 // CHECK IF THIS SHOULD BE A BUILTIN??? TODO TO DO
 /* Liretally checks if exit was typed into the line as the first command */
 
@@ -82,35 +107,3 @@ int	shell_loop(t_tools *tools)
 // 			error_exit(tools, 0);
 // 	}
 // }
-
-void	new_line(void)
-{
-	printf("\n");
-	rl_on_new_line();
-	rl_replace_line("", 0);
-	rl_redisplay();
-}
-
-void	handle_signals(int sig)
-{
-	if (sig == SIGINT)
-	{
-		new_line();
-		global_signal = SIGINT;
-	}
-	else if (sig == SIGTERM)
-		global_signal = SIGTERM;
-}
-
-void	init_sa(struct sigaction *sa)
-{
-	sa->sa_handler = handle_signals;
-	sigemptyset(&sa->sa_mask);
-	sa->sa_flags = 0;
-	if (sigaction(SIGINT, sa, NULL) == -1)
-	{
-		global_signal = SIGINT;
-		perror("sigaction");
-		exit(1);
-	}
-}
